refactor: rename extdata package to external
