+++ core/external/provider_similarsongs_test.go
@@ -1,11 +1,11 @@
-package extdata_test
+package external_test
 
 import (
 	"context"
 	"errors"
 
 	"github.com/navidrome/navidrome/core/agents"
-	. "github.com/navidrome/navidrome/core/extdata"
+	. "github.com/navidrome/navidrome/core/external"
 	"github.com/navidrome/navidrome/model"
 	"github.com/navidrome/navidrome/tests"
 	. "github.com/onsi/ginkgo/v2"
