+++ core/external/provider_updatealbuminfo_test.go
@@ -1,4 +1,4 @@
-package extdata_test
+package external_test
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 
 	"github.com/navidrome/navidrome/conf"
 	"github.com/navidrome/navidrome/core/agents"
-	"github.com/navidrome/navidrome/core/extdata"
+	"github.com/navidrome/navidrome/core/external"
 	"github.com/navidrome/navidrome/log"
 	"github.com/navidrome/navidrome/model"
 	"github.com/navidrome/navidrome/tests"
@@ -24,7 +24,7 @@ func init() {
 var _ = Describe("Provider - UpdateAlbumInfo", func() {
 	var (
 		ctx           context.Context
-		p             extdata.Provider
+		p             external.Provider
 		ds            *tests.MockDataStore
 		ag            *mockAgents
 		mockAlbumRepo *tests.MockAlbumRepo
@@ -34,7 +34,7 @@ var _ = Describe("Provider - UpdateAlbumInfo", func() {
 		ctx = GinkgoT().Context()
 		ds = new(tests.MockDataStore)
 		ag = new(mockAgents)
-		p = extdata.NewProvider(ds, ag)
+		p = external.NewProvider(ds, ag)
 		mockAlbumRepo = ds.Album(ctx).(*tests.MockAlbumRepo)
 		conf.Server.DevAlbumInfoTimeToLive = 1 * time.Hour
 	})
