+++ core/wire_providers.go
@@ -3,7 +3,7 @@ package core
 import (
 	"github.com/google/wire"
 	"github.com/navidrome/navidrome/core/agents"
-	"github.com/navidrome/navidrome/core/extdata"
+	"github.com/navidrome/navidrome/core/external"
 	"github.com/navidrome/navidrome/core/ffmpeg"
 	"github.com/navidrome/navidrome/core/metrics"
 	"github.com/navidrome/navidrome/core/playback"
@@ -18,8 +18,8 @@ var Set = wire.NewSet(
 	NewShare,
 	NewPlaylists,
 	agents.GetAgents,
-	extdata.NewProvider,
-	wire.Bind(new(extdata.Agents), new(*agents.Agents)),
+	external.NewProvider,
+	wire.Bind(new(external.Agents), new(*agents.Agents)),
 	ffmpeg.New,
 	scrobbler.GetPlayTracker,
 	playback.GetInstance,
